Revert "Add AppKit support, rename"
This reverts commit 38cea7d2a463c6fd06c85e35acf44140d67c9413.
