--- Sources/Extensions/UICollectionView+Reload.swift
@@ -1,79 +1,0 @@
-//
-//  UICollectionView+Reload.swift
-//  DifferenceKit
-//
-//  Created by Oskar Groth on 2018-09-15.
-//  Copyright © 2018 Ryo Aoyama. All rights reserved.
-//
-
-#if os(iOS) || os(tvOS)
-import UIKit
-
-public extension UICollectionView {
-    /// Applies multiple animated updates in stages using `StagedChangeset`.
-    ///
-    /// - Note: There are combination of changes that crash when applied simultaneously in `performBatchUpdates`.
-    ///         Assumes that `StagedChangeset` has a minimum staged changesets to avoid it.
-    ///         The data of the data-source needs to be updated synchronously before `performBatchUpdates` in every stages.
-    ///
-    /// - Parameters:
-    ///   - stagedChangeset: A staged set of changes.
-    ///   - interrupt: A closure that takes an changeset as its argument and returns `true` if the animated
-    ///                updates should be stopped and performed reloadData. Default is nil.
-    ///   - setData: A closure that takes the collection as a parameter.
-    ///              The collection should be set to data-source of UICollectionView.
-    func reload<C>(
-        using stagedChangeset: StagedChangeset<C>,
-        interrupt: ((Changeset<C>) -> Bool)? = nil,
-        setData: (C) -> Void
-        ) {
-        if case .none = window, let data = stagedChangeset.last?.data {
-            setData(data)
-            return reloadData()
-        }
-        
-        for changeset in stagedChangeset {
-            if let interrupt = interrupt, interrupt(changeset), let data = stagedChangeset.last?.data {
-                setData(data)
-                return reloadData()
-            }
-            
-            performBatchUpdates({
-                setData(changeset.data)
-                
-                if !changeset.sectionDeleted.isEmpty {
-                    deleteSections(IndexSet(changeset.sectionDeleted))
-                }
-                
-                if !changeset.sectionInserted.isEmpty {
-                    insertSections(IndexSet(changeset.sectionInserted))
-                }
-                
-                if !changeset.sectionUpdated.isEmpty {
-                    reloadSections(IndexSet(changeset.sectionUpdated))
-                }
-                
-                for (source, target) in changeset.sectionMoved {
-                    moveSection(source, toSection: target)
-                }
-                
-                if !changeset.elementDeleted.isEmpty {
-                    deleteItems(at: changeset.elementDeleted.map { IndexPath(item: $0.element, section: $0.section) })
-                }
-                
-                if !changeset.elementInserted.isEmpty {
-                    insertItems(at: changeset.elementInserted.map { IndexPath(item: $0.element, section: $0.section) })
-                }
-                
-                if !changeset.elementUpdated.isEmpty {
-                    reloadItems(at: changeset.elementUpdated.map { IndexPath(item: $0.element, section: $0.section) })
-                }
-                
-                for (source, target) in changeset.elementMoved {
-                    moveItem(at: IndexPath(item: source.element, section: source.section), to: IndexPath(item: target.element, section: target.section))
-                }
-            })
-        }
-    }
-}
-#endif
